Fixed minor typo -- not just making code concise
When you subtracted sequence_length, you subtracted 1 as well instead of adding 1.

If your data has a length of 10 and your window is of 2, you should append 9 items to the results array which is 10 - 2 + 1. Currently, you are appending 2 fewer items than you should: 10 - 2 - 1 = 7 for every value of len(data) and seq_len.